+++ app/sections/image-gallery/index.tsx
@@ -1,4 +1,4 @@
-import type { HydrogenComponentSchema } from "@weaverse/hydrogen";
+import { createSchema } from "@weaverse/hydrogen";
 import { forwardRef } from "react";
 import type { SectionProps } from "~/components/section";
 import { Section, sectionSettings } from "~/components/section";
@@ -16,7 +16,7 @@ let ImageGallery = forwardRef<HTMLElement, ImageGalleryProps>((props, ref) => {
 
 export default ImageGallery;
 
-export let schema: HydrogenComponentSchema = {
+export let schema = createSchema({
   type: "image-gallery",
   title: "Image gallery",
   childTypes: ["subheading", "heading", "paragraph", "image-gallery--items"],
@@ -37,4 +37,4 @@ export let schema: HydrogenComponentSchema = {
       },
     ],
   },
-};
+});