+++ CHANGELOG.md
@@ -1,5 +1,40 @@
 # @weaverse/pilot
 
+## 5.1.5
+
+### Patch Changes
+
+- Updated dependencies [43f3da3]
+  - @weaverse/hydrogen@5.2.4
+
+## 5.1.4
+
+### Patch Changes
+
+- @weaverse/hydrogen@5.2.3
+
+## 5.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @weaverse/hydrogen@5.2.2
+
+## 5.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @weaverse/hydrogen@5.2.1
+
+## 5.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @weaverse/hydrogen@5.2.0
+  - @weaverse/schema@0.4.0
+
 ## 5.0.0
 
 ### Patch Changes